backend/meal_planner/requirements.txt
Django==5.1.2
djangorestframework==3.15.2
django-cors-headers==4.5.0
django-environ==0.11.2
psycopg2-binary==2.9.9
python-dateutil==2.8.2
pytz==2024.1
requests==2.32.3
urllib3==2.2.3
PyJWT==2.8.0
python-dotenv==1.0.1
whitenoise==6.9.0
